Changing precision to FP16

+++ src/bindings/python/src/openvino/frontend/pytorch/torchdynamo/compile.py
@@ -131,6 +131,6 @@ def openvino_compile(gm: GraphModule, *args, model_hash_str: str = None):
     if model_hash_str is not None:
         core.set_property({'CACHE_DIR': cache_root + '/blob'})
     
-    device_config = {"INFERENCE_PRECISION_HINT": "f32"}
+    device_config = {"INFERENCE_PRECISION_HINT": "f16"}
     compiled = core.compile_model(om, device, device_config)
     return compiled
